feat: 채팅방 구분
